make caller responsible for creating client
I think it's a probably a bit nicer to have the caller responsible
for creating (and closing) the containerd client. It simplifies the
code and means there won't be any need to pass through creation
parameters to the registry.

This does assume that the client type is OK for concurrent use though,
so YMMV.

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>